overrides resnets with resnet1.5 version
