+++ nupic/research/frameworks/dynamic_sparse/networks/__init__.py
@@ -3,7 +3,6 @@ from .hebbian import *
 from .hebbian_v0 import GSCHeb as GSCHeb_v0
 from .layers import *
 from .main import *
-# from .resnet import *
-from .resnet_mlperf import *
 from .utils import *
-# from .wideresnet import *
+from .resnet import *
+from .wideresnet import *
